+++ src/test/converter2/issues/gh2949.js
@@ -1,0 +1,5 @@
+/**
+ * @callback Test
+ * @returns {Promise<*>}
+ */
+export {};